Add unit tests. Update the roadmap after determining testing needs.
At this point, we've written all the tests that we can easily write
by loading a single HTML fixture file (test.html) and calling
filter functions on it. For the remaining tests, we'll need the ability
to generate HTML on demand. This will allow us to consider common cases
such as suggestion threads with comments. It will also make it easier to
read fixtures, since these can be simplified HTML elements that contain
only what we want them to.