Fix navigation isssue
If you resolve a discussion matching search criteria, clicking "next" would not
take you to the expected next discussion, but instead would return you to the first
discussion. This commit introduces a fix by storing each CommentThread in memory with
an ordinal score of its position relative to other CommentThreads. This is calculated
in the constructor. We can then use this score to sort CommentThreads in NextButton and
PrevButton. Since the score is relative to all CommentThreads, it applies to both
filtered and unfiltered ThreadCollections. We can compare the score of the previously
selected (but resolved/deleted) CommentThread to the scores of CommentThreads currently
in the document to determine where to navigate.

- Add an e2e test to reproduce the bug
- Change the interface for ParseForThreads to include the discussion context element
- Remove the need to ordeer CommentThreads by their appearance on page and querying the
  index of a CommentThread within an array of CommentThreads in a NavButton target function
- simplify startingDiscussionForNavigation()